fix: deleted example dialog folder
deleted example dialog folder and moved its content to theme component
