--- src/app/shared/ui/modals/example dialog/example-dialog.component.html
@@ -1,3 +1,0 @@
-<!-- Komponent stworzony na potrzeby wyśietlenia dialogów oraz przekazania do nich danych, w wersji produkcyjnej do usuniecia -->
-      <button mat-raised-button (click)="openDialog()">Dialog z inputem</button>
-      <button mat-raised-button (click)="openDialog2()">Dialog z przyciskami</button>
