+++ src/app/shared/ui/modals/index.ts
@@ -1,0 +1,3 @@
+export * from './input-dialog/dialog-data.interface';
+export * from './input-dialog/input-dialog.component';
+export * from './quit-section-dialog/cancel-confirm-dialog.component.ts';
