+++ src/app/shared/ui/modals/input-dialog/dialog-data.interface.ts
@@ -1,0 +1,6 @@
+export interface DialogData {
+  title: string;
+  buttonText: string;
+  inputLabelText: string;
+  importedDialogData: string;
+}
