+++ src/app/shared/ui/modals/quit-section-dialog/cancel-confirm-dialog.component.html
@@ -8,7 +8,7 @@
     <p class="cancel-confirm-dialog__title">{{ data.text }}</p>
   </div>
   <div class="cancel-confirm-dialog__btn-container">
-    <button (click)="closeDialog()" mat-button class="cancel-confirm-dialog__cancel-btn error">ANULUJ</button>
-    <button mat-button class="cancel-confirm-dialog__confirm-btn success">POTWIERDŹ</button>
+    <button (click)="closeDialog()" mat-button class="cancel-confirm-dialog__cancel-btn error">anuluj</button>
+    <button mat-button class="cancel-confirm-dialog__confirm-btn success">potwierdź</button>
   </div>
 </div>
