+++ src/app/shared/ui/modals/quit-section-dialog/cancel-confirm-dialog.component.ts.ts
@@ -1,10 +1,10 @@
-import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Inject, inject } from '@angular/core';
 import { MatDialog, MatDialogModule, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
 
 @Component({
-  selector: 'app-quit-section-dialog[text]',
+  selector: 'app-cancel-confirm-dialog',
   standalone: true,
   imports: [MatDialogModule, MatIconModule, MatButtonModule],
   templateUrl: './cancel-confirm-dialog.component.html',
@@ -12,8 +12,10 @@ import { MatButtonModule } from '@angular/material/button';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class CancelConfirmDialogComponent {
-  constructor( public dialog:MatDialog, @Inject(MAT_DIALOG_DATA) public data: { text: string }) {}
+  private dialog = inject(MatDialog);
+
+  constructor(@Inject(MAT_DIALOG_DATA) public data: { text: string }) {}
   closeDialog() {
-    this.dialog.closeAll()
+    this.dialog.closeAll();
   }
 }
